mllm-demo/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mllm-demo"
version = "0.1.0"
description = "A tiny staged MLLM demo inspired by vlm-from-scratch."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch",
  "torchvision",
  "transformers>=4.40",
  "pillow",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
